# docs/phase5/logs/openapi-v1-phase5.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "DerasaX.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/Lessons/GetLessonsByUnitId": {
      "get": {
        "tags": [
          "Lessons"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Lessons/GetLessonsByUnitId": {
      "get": {
        "tags": [
          "Lessons"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Lessons/AddLesson": {
      "post": {
        "tags": [
          "Lessons"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Title": {
                    "type": "string"
                  },
                  "Content": {
                    "type": "string"
                  },
                  "UnitId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Title": {
                  "style": "form"
                },
                "Content": {
                  "style": "form"
                },
                "UnitId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Lessons/AddLesson": {
      "post": {
        "tags": [
          "Lessons"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Title": {
                    "type": "string"
                  },
                  "Content": {
                    "type": "string"
                  },
                  "UnitId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Title": {
                  "style": "form"
                },
                "Content": {
                  "style": "form"
                },
                "UnitId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Lessons/UpdateLesson": {
      "put": {
        "tags": [
          "Lessons"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Id": {
                    "type": "string"
                  },
                  "Title": {
                    "type": "string"
                  },
                  "Content": {
                    "type": "string"
                  },
                  "UnitId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Id": {
                  "style": "form"
                },
                "Title": {
                  "style": "form"
                },
                "Content": {
                  "style": "form"
                },
                "UnitId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Lessons/UpdateLesson": {
      "put": {
        "tags": [
          "Lessons"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Id": {
                    "type": "string"
                  },
                  "Title": {
                    "type": "string"
                  },
                  "Content": {
                    "type": "string"
                  },
                  "UnitId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Id": {
                  "style": "form"
                },
                "Title": {
                  "style": "form"
                },
                "Content": {
                  "style": "form"
                },
                "UnitId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Lessons/DeleteLesson": {
      "delete": {
        "tags": [
          "Lessons"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Lessons/DeleteLesson": {
      "delete": {
        "tags": [
          "Lessons"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/performance/subject/{subjectId}": {
      "get": {
        "tags": [
          "Performance"
        ],
        "parameters": [
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/student/lessons/{lessonId}": {
      "get": {
        "tags": [
          "StudentLessons"
        ],
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/Subjects/GetSubjects": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "parameters": [
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Subjects/GetSubjects": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "parameters": [
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Subjects/GetSubjectById/{id}": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Subjects/GetSubjectById/{id}": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Subjects/GetSubjectsByGradeIdAsync": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Subjects/GetSubjectsByGradeIdAsync": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Subjects/AddSubject": {
      "post": {
        "tags": [
          "Subjects"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "ImageUrl": {
                    "type": "string",
                    "format": "binary"
                  },
                  "GradeId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "ImageUrl": {
                  "style": "form"
                },
                "GradeId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Subjects/AddSubject": {
      "post": {
        "tags": [
          "Subjects"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "ImageUrl": {
                    "type": "string",
                    "format": "binary"
                  },
                  "GradeId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "ImageUrl": {
                  "style": "form"
                },
                "GradeId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Subjects/UpdateSubject": {
      "put": {
        "tags": [
          "Subjects"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Id": {
                    "type": "string"
                  },
                  "Name": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "ImageUrl": {
                    "type": "string",
                    "format": "binary"
                  },
                  "GradeId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Id": {
                  "style": "form"
                },
                "Name": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "ImageUrl": {
                  "style": "form"
                },
                "GradeId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Subjects/UpdateSubject": {
      "put": {
        "tags": [
          "Subjects"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Id": {
                    "type": "string"
                  },
                  "Name": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "ImageUrl": {
                    "type": "string",
                    "format": "binary"
                  },
                  "GradeId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Id": {
                  "style": "form"
                },
                "Name": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "ImageUrl": {
                  "style": "form"
                },
                "GradeId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Subjects/DeleteSubject/{id}": {
      "delete": {
        "tags": [
          "Subjects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Subjects/DeleteSubject/{id}": {
      "delete": {
        "tags": [
          "Subjects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}